Moved type declerations
