+++ src/web/components/ApiKeyManagement/KeyTable.stories.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable camelcase */
 import type { Meta, StoryObj } from '@storybook/react';
 
 import KeyTable from './KeyTable';
@@ -17,50 +18,50 @@ export const ManyKeys: Story = {
         contact: 'ApiKey1',
         name: 'ApiKey1',
         created: 1702860516,
-        keyId: 'PTFD_12',
+        key_id: 'PTFD_12',
         disabled: false,
         roles: [],
-        serviceId: 0,
+        service_id: 0,
       },
       {
         contact: 'ApiKey2',
         name: 'ApiKey2',
         created: 1701210253,
-        keyId: 'FDSL,089',
+        key_id: 'FDSL,089',
         disabled: false,
         roles: [{ id: 1, roleName: 'Role1', externalName: 'Role 1' }],
-        serviceId: 0,
+        service_id: 0,
       },
       {
         contact: 'ApiKey3',
         name: 'ApiKey3',
         created: 1702830516,
-        keyId: 'FDSL,089',
+        key_id: 'FDSL,089',
         disabled: false,
         roles: [
           { id: 1, roleName: 'Role1', externalName: 'Role 1' },
           { id: 2, roleName: 'Role2', externalName: 'Role 2' },
         ],
-        serviceId: 0,
+        service_id: 0,
       },
       {
         contact: 'ApiKey4',
         name: 'ApiKey4',
         created: 17028300516,
-        keyId: 'REIO_38',
+        key_id: 'REIO_38',
         disabled: false,
         roles: [
           { id: 1, roleName: 'Role1', externalName: 'Role 1' },
           { id: 2, roleName: 'Role2', externalName: 'Role 2' },
           { id: 2, roleName: 'Role3', externalName: 'Role 3' },
         ],
-        serviceId: 0,
+        service_id: 0,
       },
       {
         contact: 'ApiKey5',
         name: 'ApiKey5',
         created: 1702840516,
-        keyId: 'BNJMB_934',
+        key_id: 'BNJMB_934',
         disabled: false,
         roles: [
           { id: 1, roleName: 'Role1', externalName: 'Role 1' },
@@ -68,7 +69,7 @@ export const ManyKeys: Story = {
           { id: 2, roleName: 'Role3', externalName: 'Role 3' },
           { id: 2, roleName: 'Role3', externalName: 'Role 4' },
         ],
-        serviceId: 0,
+        service_id: 0,
       },
     ],
   },
